// src/models/types.ts
/** Plataformas de los videojuegos */
export enum Plataforma {
  PC = "PC",
  PS5 = "PlayStation 5",
  Xbox = "Xbox Series X/S",
  Switch = "Nintendo Switch",
  SteamDeck = "Steam Deck"
}

/** Posibles géneros de los videojuegos*/
export enum Genero {
  Accion = "Acción",
  Aventura = "Aventura",
  Rol = "Rol",
  Estrategia = "Estrategia",
  Deportes = "Deportes",
  Simulacion = "Simulación"
} 